requirements.txt
gradio==5.29.0
openai>=1.0.0
python-dotenv==1.0.0
requests==2.31.0
azure-ai-documentintelligence==1.0.2
azure-identity==1.15.0
openpyxl==3.1.2
pandas==2.1.4
matplotlib==3.8.2
scikit-learn
pymupdf==1.23.7
azure-ai-projects>=1.0.0b10
aiohttp>=3.8.0
markdown